Migrating from old spec to new spec
