+++ spec/spec_helper.rb
@@ -1,6 +1,7 @@
 # encoding: UTF-8
 
 require 'simplecov'
+require 'webmock/minitest'
 
 module SimpleCov
   # Setup SimpleCov.
@@ -35,7 +36,6 @@ require 'minitest/unit'
 require 'minitest/reporters'
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 require 'mocha/mini_test'
-require 'vcr_setup'
 
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
