+++ spec_old/fixtures/claims_status.json
@@ -1,0 +1,15 @@
+{
+    "patient": {
+        "birth_date": "1970-01-01",
+        "first_name": "JANE",
+        "last_name": "DOE",
+        "id": "1234567890"
+    },
+    "provider": {
+        "first_name": "Jerome",
+        "last_name": "Aya-Ay",
+        "npi": "1467560003"
+    },
+    "service_date": "2014-01-01",
+    "trading_partner_id": "MOCKPAYER"
+}
