+++ spec_old/fixtures/referrals.json
@@ -1,0 +1,36 @@
+{
+    "event": {
+        "category": "specialty_care_review",
+        "certification_type": "initial",
+        "delivery": {
+            "quantity": 1,
+            "quantity_qualifier": "visits"
+        },
+        "diagnoses": [
+            {
+                "code": "384.20",
+                "date": "2014-09-30"
+            }
+        ],
+        "place_of_service": "office",
+        "provider": {
+            "first_name": "JOHN",
+            "npi": "1154387751",
+            "last_name": "FOSTER",
+            "phone": "8645822900"
+        },
+        "type": "consultation"
+    },
+    "patient": {
+        "birth_date": "1970-01-01",
+        "first_name": "JANE",
+        "last_name": "DOE",
+        "id": "1234567890"
+    },
+    "provider": {
+        "first_name": "CHRISTINA",
+        "last_name": "BERTOLAMI",
+        "npi": "1619131232"
+    },
+    "trading_partner_id": "MOCKPAYER"
+}
