+++ spec_old/fixtures/sample.270
@@ -1,0 +1,131 @@
+ISA*00*          *00*          *30*SENDER         *30*RECEIVER       *140305*1953*^*00501*000000905*0*T*:~
+GS*HS*SENDER_ID*RECEIVER_ID*20140305*1953*CONTROL_ID*X*005010X279A1~
+ST*270*0001*005010X279A1~
+BHT*0022*13*813785794*20140305*1953~
+HL*1**20*1~
+NM1*PR*2*MOCKPAYER*****PI*MOCKPAYER~
+HL*2*1*21*1~
+NM1*1P*2*Wuckert Inc*****XX*581603684~
+HL*3*2*22*0~
+TRN*1*181399807*9324422845~
+NM1*IL*1*Tromp*Jerod****MI*545943979~
+DMG*D8*19960507~
+DTP*291*D8*20140305~
+EQ*30~
+HL*4*2*22*0~
+TRN*1*611822238*9324422845~
+NM1*IL*1*Walker*Hassan****MI*773415451~
+DMG*D8*19960613~
+DTP*291*D8*20140305~
+EQ*30~
+HL*5*2*22*0~
+TRN*1*530994264*9324422845~
+NM1*IL*1*Johns*Cathrine****MI*505837362~
+DMG*D8*19881001~
+DTP*291*D8*20140305~
+EQ*30~
+HL*6*2*22*0~
+TRN*1*145769779*9324422845~
+NM1*IL*1*Durgan*Stephon****MI*204791243~
+DMG*D8*19830509~
+DTP*291*D8*20140305~
+EQ*30~
+HL*7*2*22*0~
+TRN*1*853276515*9324422845~
+NM1*IL*1*Erdman*Bartholome****MI*994176438~
+DMG*D8*19930513~
+DTP*291*D8*20140305~
+EQ*30~
+HL*8*2*22*0~
+TRN*1*392506326*9324422845~
+NM1*IL*1*Padberg*Kaci****MI*331914222~
+DMG*D8*19920730~
+DTP*291*D8*20140305~
+EQ*30~
+HL*9*2*22*0~
+TRN*1*669106647*9324422845~
+NM1*IL*1*Donnelly*Kyleigh****MI*461763996~
+DMG*D8*19890712~
+DTP*291*D8*20140305~
+EQ*30~
+HL*10*2*22*0~
+TRN*1*915888295*9324422845~
+NM1*IL*1*Monahan*Jimmie****MI*315531994~
+DMG*D8*19950414~
+DTP*291*D8*20140305~
+EQ*30~
+HL*11*2*22*0~
+TRN*1*924598474*9324422845~
+NM1*IL*1*Bednar*Markus****MI*674957204~
+DMG*D8*19780109~
+DTP*291*D8*20140305~
+EQ*30~
+HL*12*2*22*0~
+TRN*1*218563674*9324422845~
+NM1*IL*1*McCullough*Leland****MI*789318226~
+DMG*D8*19800202~
+DTP*291*D8*20140305~
+EQ*30~
+HL*13*2*22*0~
+TRN*1*764225581*9324422845~
+NM1*IL*1*Macejkovic*Norwood****MI*605486216~
+DMG*D8*19890726~
+DTP*291*D8*20140305~
+EQ*30~
+HL*14*2*22*0~
+TRN*1*843675248*9324422845~
+NM1*IL*1*Ortiz*Keyon****MI*178470414~
+DMG*D8*19920126~
+DTP*291*D8*20140305~
+EQ*30~
+HL*15*2*22*0~
+TRN*1*575530276*9324422845~
+NM1*IL*1*Doyle*Olin****MI*727454392~
+DMG*D8*19890922~
+DTP*291*D8*20140305~
+EQ*30~
+HL*16*2*22*0~
+TRN*1*583253119*9324422845~
+NM1*IL*1*Doyle*Greg****MI*302952193~
+DMG*D8*19890228~
+DTP*291*D8*20140305~
+EQ*30~
+HL*17*2*22*0~
+TRN*1*494693024*9324422845~
+NM1*IL*1*Brekke*Nathanial****MI*228065686~
+DMG*D8*19910131~
+DTP*291*D8*20140305~
+EQ*30~
+HL*18*2*22*0~
+TRN*1*654387679*9324422845~
+NM1*IL*1*Herzog*Jayme****MI*653618078~
+DMG*D8*19871017~
+DTP*291*D8*20140305~
+EQ*30~
+HL*19*2*22*0~
+TRN*1*864981376*9324422845~
+NM1*IL*1*Ebert*Heaven****MI*548879156~
+DMG*D8*19930416~
+DTP*291*D8*20140305~
+EQ*30~
+HL*20*2*22*0~
+TRN*1*618055454*9324422845~
+NM1*IL*1*Cummerata*Hector****MI*688306042~
+DMG*D8*19780613~
+DTP*291*D8*20140305~
+EQ*30~
+HL*21*2*22*0~
+TRN*1*608414882*9324422845~
+NM1*IL*1*Kerluke*Enola****MI*687416329~
+DMG*D8*19960902~
+DTP*291*D8*20140305~
+EQ*30~
+HL*22*2*22*0~
+TRN*1*568364627*9324422845~
+NM1*IL*1*Wehner*Alyce****MI*728313788~
+DMG*D8*19701130~
+DTP*291*D8*20140305~
+EQ*30~
+SE*127*0001~
+GE*1*CONTROL_ID~
+IEA*1*000000905~
