+++ spec_old/spec_helper.rb
@@ -1,9 +1,6 @@
 # encoding: UTF-8
 
 require 'simplecov'
-require 'webmock/minitest'
-
-# WebMock.disable_net_connect!
 
 module SimpleCov
   # Setup SimpleCov.
@@ -38,6 +35,7 @@ require 'minitest/unit'
 require 'minitest/reporters'
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 require 'mocha/mini_test'
+require 'vcr_setup'
 
 $LOAD_PATH.unshift(File.dirname(__FILE__))
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..', 'lib'))
